+++ src/main/java/pe/com/tatapizzeria/repository/ComprobantesPagoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import pe.com.tatapizzeria.entity.ComprobantesPagoEntity;
 import java.util.List;
 
-public interface ComprobantesPagoRepository extends JpaRepository<ComprobantesPagoEntity, Long> {
+public interface ComprobantesPagoRepository extends JpaRepository<ComprobantesPagoEntity, Integer> {
 
     @Query("SELECT s FROM ComprobantesPagoEntity s WHERE s.estado = true")
     List<ComprobantesPagoEntity> findAllCustom();
