+++ src/main/java/pe/com/tatapizzeria/repository/PedidosRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import pe.com.tatapizzeria.entity.PedidosEntity;
 import java.util.List;
 
-public interface PedidosRepository extends JpaRepository<PedidosEntity, Long> {
+public interface PedidosRepository extends JpaRepository<PedidosEntity, Integer> {
 
     @Query("SELECT s FROM PedidosEntity s WHERE s.estado = true")
     List<PedidosEntity> findAllCustom();
