+++ src/main/java/pe/com/tatapizzeria/repository/RepartidoresRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import pe.com.tatapizzeria.entity.RepartidoresEntity;
 import java.util.List;
 
-public interface RepartidoresRepository extends JpaRepository<RepartidoresEntity, Long> {
+public interface RepartidoresRepository extends JpaRepository<RepartidoresEntity, Integer> {
 
     @Query("SELECT s FROM RepartidoresEntity s WHERE s.estado = true")
     List<RepartidoresEntity> findAllCustom();
