+++ src/main/java/pe/com/tatapizzeria/repository/SedesRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import pe.com.tatapizzeria.entity.SedesEntity;
 import java.util.List;
 
-public interface SedesRepository extends JpaRepository<SedesEntity, Long> {
+public interface SedesRepository extends JpaRepository<SedesEntity, Integer> {
 
     @Query("SELECT s FROM SedesEntity s WHERE s.estado = true")
     List<SedesEntity> findAllCustom();
