+++ src/main/java/pe/com/tatapizzeria/restcontroller/ComprobantesPagoRestController.java
@@ -19,17 +19,17 @@ public class ComprobantesPagoRestController {
     public List<ComprobantesPagoEntity> findAllCustom() { return s.findAllCustom(); }
 
     @GetMapping("/{id}")
-    public ComprobantesPagoEntity findById(@PathVariable Long id) { return s.findById(id); }
+    public ComprobantesPagoEntity findById(@PathVariable Integer id) { return s.findById(id); }
 
     @PostMapping
     public ComprobantesPagoEntity add(@RequestBody ComprobantesPagoEntity o) { return s.add(o); }
 
     @PutMapping("/{id}")
-    public ComprobantesPagoEntity update(@RequestBody ComprobantesPagoEntity o, @PathVariable Long id) { return s.update(o, id); }
+    public ComprobantesPagoEntity update(@RequestBody ComprobantesPagoEntity o, @PathVariable Integer id) { return s.update(o, id); }
 
     @DeleteMapping("/{id}")
-    public ComprobantesPagoEntity delete(@PathVariable Long id) { return s.delete(id); }
+    public ComprobantesPagoEntity delete(@PathVariable Integer id) { return s.delete(id); }
 
     @PatchMapping("/{id}")
-    public ComprobantesPagoEntity enable(@PathVariable Long id) { return s.enable(id); }
+    public ComprobantesPagoEntity enable(@PathVariable Integer id) { return s.enable(id); }
 }
