+++ src/main/java/pe/com/tatapizzeria/restcontroller/TamanosRestController.java
@@ -19,17 +19,17 @@ public class TamanosRestController {
     public List<TamanosEntity> findAllCustom() { return s.findAllCustom(); }
 
     @GetMapping("/{id}")
-    public TamanosEntity findById(@PathVariable Long id) { return s.findById(id); }
+    public TamanosEntity findById(@PathVariable Integer id) { return s.findById(id); }
 
     @PostMapping
     public TamanosEntity add(@RequestBody TamanosEntity o) { return s.add(o); }
 
     @PutMapping("/{id}")
-    public TamanosEntity update(@RequestBody TamanosEntity o, @PathVariable Long id) { return s.update(o, id); }
+    public TamanosEntity update(@RequestBody TamanosEntity o, @PathVariable Integer id) { return s.update(o, id); }
 
     @DeleteMapping("/{id}")
-    public TamanosEntity delete(@PathVariable Long id) { return s.delete(id); }
+    public TamanosEntity delete(@PathVariable Integer id) { return s.delete(id); }
 
     @PatchMapping("/{id}")
-    public TamanosEntity enable(@PathVariable Long id) { return s.enable(id); }
+    public TamanosEntity enable(@PathVariable Integer id) { return s.enable(id); }
 }
