+++ src/main/java/pe/com/tatapizzeria/service/ClientesService.java
@@ -6,9 +6,9 @@ import pe.com.tatapizzeria.entity.ClientesEntity;
 public interface ClientesService {
     List<ClientesEntity> findAll();
     List<ClientesEntity> findAllCustom();
-    ClientesEntity findById(Long id);
+    ClientesEntity findById(Integer id);
     ClientesEntity add(ClientesEntity obj);
-    ClientesEntity update(ClientesEntity obj, Long id);
-    ClientesEntity delete(Long id);
-    ClientesEntity enable(Long id);
+    ClientesEntity update(ClientesEntity obj,Integer id);
+    ClientesEntity delete(Integer id);
+    ClientesEntity enable(Integer id);
 }
