+++ src/main/java/pe/com/tatapizzeria/service/HistorialEstadosPedidoService.java
@@ -5,14 +5,14 @@ import pe.com.tatapizzeria.entity.HistorialEstadosPedidoEntity;
 
 public interface HistorialEstadosPedidoService {
     List<HistorialEstadosPedidoEntity> findAll();
-    HistorialEstadosPedidoEntity findById(Long id);
+    HistorialEstadosPedidoEntity findById(Integer id);
     HistorialEstadosPedidoEntity add(HistorialEstadosPedidoEntity obj);
     
     // Obtener TODOS los historiales de un pedido
-    List<HistorialEstadosPedidoEntity> findByPedidoId(Long idPedido);
+    List<HistorialEstadosPedidoEntity> findByPedidoId(Integer idPedido);
     
     // Obtener el último estado de un pedido
-    HistorialEstadosPedidoEntity findLastEstadoByPedidoId(Long idPedido);
+    HistorialEstadosPedidoEntity findLastEstadoByPedidoId(Integer idPedido);
     
     // Obtener el último estado de cada pedido
     List<HistorialEstadosPedidoEntity> findLastEstadoForEachPedido();
